requirements.txt
pytest
selenium
pyzmail36
python-dotenv
openpyxl
imapclient
pytest-html